an acceptable Crd solution
